backend/requirements.txt
fastapi
uvicorn
pydantic
pypdf
tiktoken
faiss-cpu
numpy
requests
reportlab
python-dotenv
groq
openai
sentence-transformers
jsonschema